HYD-6908 Remove unnecessary fetchhelp's (#32)
Chris Gearing has tended to not remove fetchhelps (through laziness) when
landing code. This means I get messages for failures that are really just
test failures and not related to the message. He has added them to a junk
mail filter!

But when this happens tests take additional time.

To fix this double whammy burger I've removed the fetch helps for
chris.gearing@intel.com and left just the one for HYD-4050 which goes to
Brian.

  - Remove 5 fetchelp usages
  - Change usage documentation to include names of current developers

Change-Id: I600ce5cde71dd27fa2c007d7266a3df3ceb988bd
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>