Remove Downloading of npm during build.
We are currently using a bridged version of node
sans npm.

We were using this code to pull npm for builds, however
npm has been added to the builder, which obviates this.

Remove npm and the code that setup a new npm root.
Since we are staying current with node / npm there
is no need to jail a newer npm setup.

Signed-off-by: Joe Grund <joe.grund@intel.com>