Fixes needed to pass SSI on the simulator
Move packages from e2fsprogs bundle to the external bundle since
the e2fsprogs bundle doesn't exist any more.

Wait for the alert to be lowered instead of expecting it to be.

In this world of fake-ness there is no session restarting so only
sending package scan updates on session startup (like we do with
the real agent) doesn't really allow for the manager to notice
package scan changes, so in this world of fakeness, send a package
scan on every session update as they are cheap to do here.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>