Add Health Status stored procedure
We ask for status and count of alerts to represent
system health.

This call currently involves

  1. Selecting every active row with a certain severity.
  2. Reducing the rows into a count + maximum severity.

Instead of this, we can use a stored procedure. This will
centralize the information for any consumer, and allows us
to get the status with a simple call: `select * from health_status()`.

Then, we can couple this call with postgres listen + notify.

This will end up providing a reactive event flow, all in the postgres
level.

That means that a consumer can subscribe to changes from the
DB. This approach will scale as well.

Signed-off-by: Joe Grund <joe.grund@intel.com>