Add Notify triggers for relevant chroma_core_* tables.
Add triggers for broadcasting Insert | Update | Delete changes on all
relevant chroma_core_* tables.

This will use Postgres NOTIFY to send a comma separated list of
operation name and id.

We should be using `json_build_object` but that's not available till
postgres 9.3.

Building off of this, any process with a client connection can do a
`LISTEN table_update` and will be asynchronously notified of changes on
the table.

Signed-off-by: Joe Grund <joe.grund@intel.com>