Remove django-supervisor
We have a dependency on django-supervisor at dev time. It has some quirks
that makes having it more difficult than not.

Particularlly, it requires a username / pw combo to use, and there
is no way to disable it.

Ironically, we do not require a un/pw to access supervisor in production.
This means we have to write special auth code for development and then
make sure to ignore it in production.

  - Remove django-supervisor.

Signed-off-by: Joe Grund <joe.grund@intel.com>