Add support for API key auth
As we move towards a distributed model, we need a way to authenticate against the api besides username / password combo.

 We currently proxy the session forward when needed, but this is sub-optimal.

This change is sub-optimal as well IMO (we should be using some sort of API gateway for auth and not concern ourselves with it in this service), but it enables further changes.