Use the lustre-zfs-ldiskfs meta-package
To define what should be installed for a managed file system profile.

But in doing so, we remove the list of packages from the profile and
this will break upgrade handling so we will refactor upgrade handling
(at least) in an upcoming patch.

This meta-package also gives us a place to work around LU-9745 until
that is fixed upstream.