Create needed repos during host deployment
Have the agent deployment mechanism create the required repos for
installing the agent's dependencies instead of requiring that the
storage servers have these repos pre-configured.

This requires LU-9776 to refer to the client RPM as
patchless-lustre-client and as such we are using the Copr
lustre-client repository temporarily.  If LU-9776 is not accepted
by the Lustre team, we will have to refactor this support into
separate repos for the lustre server and client nodes, unfortunately
increasing the complexity for us.

For testing with Vagrant, we need to prefer IPv4 over IPv6 in DNS
results since the cluster won't likely have IPv6.

A good client should automatically fall-back from an unusable
IPv6 AAAA result to an IPv4 A result but not all clients are
good.

So given that the Vagrant nodes could have IPv6 configuration
but not IPv6 routing, just prefer A results over AAAA.

Since we are installing the EPEL repo, we don't really need to then also
check for it.

Arguably, the installation of it should be conditional on it not being
present already.

Use build.whamcloud.com name due to conflicts with proxy configuration
when running on internal networks.

Need to add repos for EFS in cluster_setup.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>