Create needed repos during host deployment
Have the agent deployment mechanism create the required repos for
installing the agent's dependencies instead of requiring that the
storage servers have these repos pre-configured.

This requires LU-9776 to refer to the client RPM as
patchless-lustre-client and as such we are using the Copr
lustre-client repository temporarily.  If LU-9776 is not accepted
by the Lustre team, we will have to refactor this support into
separate repos for the lustre server and client nodes, unfortunately
increasing the complexity for us.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>