Update storage screen / remove highcharts
Highcharts is a paid piece of code. The major
last place it was being used is the storage screen.

This patch bumps the versions of GUI and old-gui
which have the new storage screen, and remove the
old storage screen / highcharts respectively.

In addition, remove the highcharts license, and
notice.

Signed-off-by: Joe Grund <joe.grund@intel.com>