Remove chroma-diagnostics, replace with sosreport plugin
Fixes #263

We currently have dedicated code to collect diagnostic reports, chroma-diagnostics.

This code reinvents much of the wheel that is provided by sosreport + an iml plugin.

Remove chroma-diagnostics, replace with a repo that hosts the iml-sosreport-plugin.

This repo can be published to pypi + added to copr and we can remove all chroma-diagnostics code.