Allow the user to override repo locations (#277)
* Remove superfluous whitespaces

* Allow the user to override repo locations

This is done by first setting repo locations in chroma_support.repo
in the installation directory.  This file will NOT override an
existing /etc/yum.repos.d/chroma_support.repo file however.

Then once the manager is installed, storage server repo overrides
can be made to /usr/share/chroma-manager/storage_server.repo.  This
file is marked as a configuration file in the RPM so upgrades won't
override changes to this file.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>

* Update hardcode

Update the hardcode of fencing agents, with a new hardcode.