Issue a filesystem stop in tearDown()
During the ApiTestCaseWithTestReset.setUp() process
remote_operations.clear_ha() is called to clear the Pacemaker
configuration and unmount Lustre targets in this order.

The process of clearing Pacemaker was to stop Pacemaker on each
server, in a serial fashion.  The result of stopping Pacemaker
on the first node of a cluster however meant failing over that
node's targets to the remaining node in the cluster.  Pacemaker
was then stopped on the second node.

In the process of failing over targets on the first node, the
device was unmounted and then mounted on the second node.
However if the MGS had been stopped before OSSes then OSTs
starting up as the second phase of that failover step could take
longer than Pacemaker's 300s limit and end up staring after
Pacemaker (and our test framework) thought it was stopped.
See LU-9925 for more details.

To mitigate all of this, add an orderly stopping of all filesystems
to ApiTestCaseWithTestReset.tearDown() so that targets are all just
shutdown and not failed over, prior to stopping Pacemaker (which
still happens in setUp() but arguably should be part of tearDown()
also).

Also, error check the unmounting of lustre targets and the clear_ha
script execution.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>