Wait for udev to settle
partprobe causes partitions to thrash back and forth.

If we don't wait for udev to settle, then we may get the devices in the middle of thrashing.

Longer term, we should push instead of pull and stream changes to the manager instead of polling.

Also revert ndp, as the settle may fix the issues.