Websockets blocked by invalid certs
Chrome has turned the security screws again, and we need to update our certs.

We are missing [SAN](https://geekflare.com/san-ssl-certificate/), and it appears even after adding it, we are still blocked by the self-signed cert.

As a first step, update our crypto generation to add a DNS name equivalent to the current CN.