Iml common bump (#319)
* pin iml-common package versions to 1.0

* pin iml-common version to 1.0.7 to fix unit tests

* pin iml-common range so we upgrade to latest patch version

* Move iml-common requirement out of requirements.dev

This belongs as a "product" requirement so should go into
chroma-dependencies/Makefile.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>

* Fix the order of stopping pacemaker and friends

We should stop the services in the order of their dependency
on each other.  Don't pull the rug out from under pacemaker
by stopping corosync first for example.

Tighten up some unquoted variables.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>

* Handle a NOOP state transition request

Fixes HYD-3468

When trying to stop a filesystem that has a worker configured but
not started one can get an error such as:

[2014-08-19 18:35:56,694: ERROR/django.request.tastypie] Internal Server Error: /api/filesystem/1/
Traceback (most recent call last):
  File "/usr/lib/python2.6/site-packages/tastypie/resources.py", line 192, in wrapper
    response = callback(request, *args, **kwargs)
  File "/usr/lib/python2.6/site-packages/tastypie/resources.py", line 406, in dispatch_detail
    return self.dispatch('detail', request, **kwargs)
  File "/usr/lib/python2.6/site-packages/tastypie/resources.py", line 427, in dispatch
    response = method(request, **kwargs)
  File "/usr/lib/python2.6/site-packages/tastypie/resources.py", line 1131, in put_detail
    updated_bundle = self.obj_update(bundle, request=request, **self.remove_api_resource_names(kwargs))
  File "/usr/share/chroma-manager/chroma_api/utils.py", line 319, in obj_update
  File "/usr/share/chroma-manager/chroma_api/utils.py", line 267, in obj_update
  File "/usr/share/chroma-manager/chroma_core/services/job_scheduler/job_scheduler_client.py", line 203, in get_transition_consequences
  File "/usr/share/chroma-manager/chroma_core/services/job_scheduler/command_plan.py", line 176, in get_transition_consequences
  File "/usr/share/chroma-manager/chroma_core/services/job_scheduler/command_plan.py", line 377, in _emit_transition_deps
  File "/usr/share/chroma-manager/chroma_core/services/job_scheduler/command_plan.py", line 456, in _collect_dependencies
  File "/usr/share/chroma-manager/chroma_core/services/job_scheduler/command_plan.py", line 369, in _emit_transition_deps
  File "/usr/share/chroma-manager/chroma_core/models/jobs.py", line 259, in get_route
SchedulingError: unmounted->unmounted not legal state transition for <class 'chroma_core.models.client_mount.LustreClientMount'>

This appears to be because our state engine does not handle NOOP state
transition requests.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>

* Remove errantly duplicated line

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>

* Issue a filesystem stop in tearDown()

During the ApiTestCaseWithTestReset.setUp() process
remote_operations.clear_ha() is called to clear the Pacemaker
configuration and unmount Lustre targets in this order.

The process of clearing Pacemaker was to stop Pacemaker on each
server, in a serial fashion.  The result of stopping Pacemaker
on the first node of a cluster however meant failing over that
node's targets to the remaining node in the cluster.  Pacemaker
was then stopped on the second node.

In the process of failing over targets on the first node, the
device was unmounted and then mounted on the second node.
However if the MGS had been stopped before OSSes then OSTs
starting up as the second phase of that failover step could take
longer than Pacemaker's 300s limit and end up staring after
Pacemaker (and our test framework) thought it was stopped.
See LU-9925 for more details.

To mitigate all of this, add an orderly stopping of all filesystems
to ApiTestCaseWithTestReset.tearDown() so that targets are all just
shutdown and not failed over, prior to stopping Pacemaker (which
still happens in setUp() but arguably should be part of tearDown()
also).

Also, error check the unmounting of lustre targets and the clear_ha
script execution.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>

* Bump iml-common to 1.2

* Update zfs.py

* Bump deps to 1.3