Add browser conf
Background

In HYD-6955, we saw that Chrome will be removing support for SHA-1.
These types of changes are likely in the browser space and we may not need to update the agent crypto for all of them.
Additionally, some sites may wish to have a non self-signed cert. Currently it's difficult to change certs for a site that has deployed their agents.

Solution
Implement a separate conf for the browser. This conf will take it's own server_name. By doing this, we can serve different certs for the browser and the rest of the app.