Read lightweight tags for release
Our code for reading tags assumes an annotated tag.

This is sub-optimal for releasing software using Github.

A typical software release on Github you follows this process.

1. Go to the releases tab on the project and create a release (this creates a lightweight tag in the process).
2. Your build system (Travis etc...) will pick up on a the release, and build the software. It will then publish it to the relevant spot (registry etc...).

We currently do not follow this workflow. Our build script is expecting an annotated tag, which means we need to:

1. Tag locally.
2. Push the tag to origin
3. Create a release from that existing tag.
4. Trigger and mark a build by hand.
5. Upload the artifact back to the release page.

We should additionally add a job that follows what Typical cloud CI providers do, but for now just allowing lightweight tags means we can create the release on github which is better than what we have to do today.