Remove the agent module as it's separate now.
When updating, we need to upgrade to python2-iml-agent as well as
create the basic minimum repo file to install on the agent during
the upgrade to complete the three stage update.

Since the packaging handling mechanism in pre-4.1 releases
did not know how to handle package renames, we have to let
the smarter DNF handle it on upgrades through 4.1.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>