Fail fast on plugin load error
The agent currently logs any exceptions during plugin load and chugs
forward. When this occurs, plugins are dropped, leading to difficult
to debug errors later on when that plugin is invoked.

Instead, we should be explicit and fail fast on a plugin import error.
Systemd will restart the unit after fail to give loading another shot.

Signed-off-by: Joe Grund <joe.grund@intel.com>