Fix LNet Detection
We are currently determining if LNet is loaded by looking for "lnet"
inside of "/proc/modules." Furthermore, we are checking to see if lnet
is up by looking at "/proc/sys/lnet/stats". This is a problem for a
couple of reasons:
1. We shouldn't have to rely on looking at these files
2. What if the files change (and they are in an upcoming version of
Lustre).

A better solution is to:
1. use the result of "udevadm info --path /sys/module/lnet" to determine if lnet is loaded
2. Use the result of "lnetctl net show" to determine if lnet is started

By doing this, we are using reliable commands and are no longer coupled
to these files.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>