Add Date serializer to log api.
Currently, the dates on the log page are being returned in the following
format:

2018-07-02T12:30:00.0000

This is a UTC timestamp but it does not include the 'Z' at the end to
indicate that it is UTC. In this case, the time is 8:30:00 EST, but the
browser will treat this ast 12:30 local time because the timezone is not
provided. To fix this, use the date serializer on the log api such that
'Z' is appended to the end of the timestamps.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>