Move installation steps to service_config
Fixes #690

Now that IML is available in docker, we need to move the config (much of it systemd specific) into chroma-config setup.

The reason this is needed is that when the RPM is installed in the docker build step, systemd is not available, but is available once someone starts the container.

Once this is done, starting the container and running chroma-config setup will yield a fully functional IML manager.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>