Make manager upgradable
Fixes #682

The pre-4.1 version of the manager is not currently upgradable to the 4.1 version.

This is because the 4.1 version is a package rename, but we haven't provided / obsoleted over the previous version.

In addition, removing a previous(ly named) version clears out existing certs, so we will need to use triggers to keep the certs alive.

Finally, workers are already set to auto in the nginx conf, we don't need to do so in the spec.

Signed-off-by: Nathaniel Clark <nclark@whamcloud.com>