Build docker images in travis
We are currently using docker cloud
to build images upon landing something in the IML repo.

This currently has two downsides:

1. manager-service-base needs to be built / pushed before everything
else builds, but there is not a way to sequence that.
2. There are very long queueing times in docker cloud, leading to large
lags between landing some changes and it moving to production.

By moving the continuous integration step to travis we alieviate both of
these issues.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>