IML container tuning (#722)
Upon trying IML in docker on a larger system, I noticed a few areas of
improvement.

   1. We should scale down our gunicorn workers a bit (max out at 8).
This improves memory usage considerably while still maintaining the same
level of responsiveness.
  2. Add a retry to intermittent queue connection issues. While
debugging
https://success.docker.com/article/ipvs-connection-timeout-issue I
noticed that kombu (rabbitMQ client lib) has a retry option. I've added
this to at least notify that we have an issue and might retry.
  3. The RabbitMQ container can be ephemeral; there is nothing durable
that needs to be stored. In addition we can use the management version
which will allow us to view connections / queues / exchanges etc... over
port 15672.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>