Create iml-settings file using atomic operation
The iml-settings.conf file was being created using redirection, which is not
an atomic operation. The result of doing this is that the file will be
created without contents. This ultimately leads to multiple docker
containers thinking environment variables are available when they
actually aren't and causes the process running in the container to
crash.

This patch changes how the iml-settings.conf file is created by first
redirecting into a temp file and then using the `mv` command to move it
to /var/lib/chroma/iml-settings.conf. The mv command is atomic and
guarantees that the contents will be available when the file exists.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>