Bump Django to 1.6
Update Django from 1.4.5 to 1.6.11.

This is a very large patch out of necessity due to numerous breaking
changes post 1.4.

The biggest change is transaction management has been overhauled.
This is actually a "good thing"; autocommit per statement is the default
and transactions are explicit, not open for the duration of a http
request.

Atomicity is also strongly enforced within transactions.

Read more about the change [here](https://realpython.com/transaction-management-with-django-1-6/).

Because of this change, I've dropped all manual commits throughout the
manager. autocommit should enforce statement level commits, and if
autocommit is off, we are in an atomic transaction that can't be
committed until the transaction closes.

DSE has been replaced with massiviu, which I had to custom build and
host in Copr to work around a packaging bug in their code. The bug has
been fixed via a PR I sent, but a proper version still needs to be released.

I also needed to update the gevent config to account for threading
issues with db connections.

In addition, I added Warning output to the unit tests and fixed all the
warnings (there were lots) in the IML code. Some warnings still remain
in 3rd party code.

There were also many imports and function / method calls that slightly
changed and needed updating.

Files that were parsable by Python3 were also formatted by black.

Needless to say, since this patch touches so much of the core, it should
also get a thorough manual review.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>