Bump Django to 1.6 (#777)
* Bump Django to 1.6

Update Django from 1.4.5 to 1.6.11.

This is a very large patch out of necessity due to numerous breaking
changes post 1.4.

The biggest change is transaction management has been overhauled.
This is actually a "good thing"; autocommit per statement is the default
and transactions are explicit, not open for the duration of a http
request.

Atomicity is also strongly enforced within transactions.

Read more about the change [here](https://realpython.com/transaction-management-with-django-1-6/).

Because of this change, I've dropped all manual commits throughout the
manager. autocommit should enforce statement level commits, and if
autocommit is off, we are in an atomic transaction that can't be
committed until the transaction closes.

DSE has been replaced with massiviu, which I had to custom build and
host in Copr to work around a packaging bug in their code. The bug has
been fixed via a PR I sent, but a proper version still needs to be released.

I also needed to update the gevent config to account for threading
issues with db connections.

In addition, I added Warning output to the unit tests and fixed all the
warnings (there were lots) in the IML code. Some warnings still remain
in 3rd party code.

There were also many imports and function / method calls that slightly
changed and needed updating.

Files that were parsable by Python3 were also formatted by black.

Needless to say, since this patch touches so much of the core, it should
also get a thorough manual review.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* - Proxy doesn't exist yet. Change the bind value in wsgi to use the old
bind value.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* Remove supervisord / replace with systemd units (#630)

Fixes #628

We are currently using supervisord for some manager service management.

Now that systemd is ubiquitous, we don't have a need for supervisord.

Further, we can tie all dependencies together and ensure reliable service ordering.

Finally, we can auto-populate env settings via a django management command
that get's run by a oneshot systemd unit to enable
various services to easily access shared config.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Add more service ordering (#639)

We need to make sure that when starting the
iml-manager.target, we both require the device-aggregator
and start wait for postgres + rabbitmq to be up.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Remove ui modules (#781)

* Remove supervisord / replace with systemd units (#630)

Fixes #628

We are currently using supervisord for some manager service management.

Now that systemd is ubiquitous, we don't have a need for supervisord.

Further, we can tie all dependencies together and ensure reliable service ordering.

Finally, we can auto-populate env settings via a django management command
that get's run by a oneshot systemd unit to enable
various services to easily access shared config.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Add more service ordering (#639)

We need to make sure that when starting the
iml-manager.target, we both require the device-aggregator
and start wait for postgres + rabbitmq to be up.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Remove ui-modules

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* namespace dir correctly

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* fixup travis test paths.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* bump requirements up

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* bump deps and change dirs

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* fixup import

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* update nginx config

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* fixup runner

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* more fixups

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* fixups

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* fixup nginx settings

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* Remove worker_process changes. (#689)

Nginx in EPEL already default to worker_process auto.

We don't need to alter it anymore in the spec.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* remove nginx.conf.template

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* Ensure redirect happens from port http to https (#664)

Signed-off-by: Nathaniel Clark <nclark@whamcloud.com>

* add .flake8

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* more fixups

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* More fixups

* - fixups

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>