Create /action endpoint
Currently for each stateful object requested via the API,
we tack on extra information regarding it's available jobs +
transitions, and locks.

This is done in:

https://github.com/whamcloud/integrated-manager-for-lustre/blob/5f8dcc3615388527c833df348617119352f51d07/chroma_api/utils.py#L198-L280

This is primarily done for convenience, but adds a bunch of overhead
both in RPC calls and database access.

Add a /action endpoint that can be called on demand, and remove the actions being appended to utils.

This will make any API polling calls for stateful objects cheaper, as additional queue / db traffic is mitigated.

To call the endpoint, do something like the following:

```
https://localhost:8443/api/action/?composite_ids=62:1&composite_ids=62:2
```

Where 62 is the content_type_id, and 1,2 are the resource ids.

The endpoint can be called to resolve 1-N records at once.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>