Implement full IML plugin stack
This patch implements the full iml plugin stack including the agent and
IML manager sides.

Both the agent and manager sides are implemented here for ease of
maintenance.

This still avoids a monolithic situation, as separate areas of
responsibility are organized into Crates under a single
[workspace](https://doc.rust-lang.org/book/ch14-03-cargo-workspaces.html).

This is now more akin to a modulith, where module specific concerns and
boundaries are maintained under a single repo.

All networked components are implemented with [Tokio](https://tokio.rs/)
to address any speed / scaling / code maintainability concerns.

  - Add Makefile + spec to build the whole cargo workspace.
  - Add section to `chroma-manager.conf` to consume agent messages in the
manager
  - Add RabbitMQ settings to shared IML manager environment.
  - Add Crate for manager-side services.
  - Add Crate to send / receive agent messages to IML manager.
  - Add Crate to run daemon / action plugins on IML agents.
  - Add Crate for proper CLI capabilities on IML agent.
  - Add a few simple reference plugins to iml-agent and iml-agent-comms crates.
  - Add a Crate to cleanly shutdown tokio runtimes on error.
  - Some small Python fixups.

Create an iml-comms consumer that somewhat mirrors chroma_agent_comms, but is in Rust.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>