Fix hanging test
An action_runner integration test was intermittently hanging.

I think this is because I was using a while loop which didn't yield long
enough for another thread to gain access to the mutex and place a value
in.

Use a loop_fn instead, with a delay of 10 milliseconds to yield the mutex back.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>