flush_queue should take a oneshot channel for cancelation
Fixes #878.

If a session is cancelled, an older session may sit draining on the flush_queue up until timeout.
Have the flush_queue take a oneshot channel that will resolve when a session is terminated.
The other half of the channel should be held by a session, and should fire on session termination.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>