Bump seed to latest version
Update seed to the latest version.

There are numerous breaking changes in this version (0.3.6)
the largest of which is changes to the fetch module which necessitates
using fetch from within the update fn.

This means we need a large refactor of the action dropdown and it's
fetch polling.

While were in here, try to extract the hsm dropdown to a more separate
component as there does not appear to be much overlas there does not
appear to be much overlap.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>