Fixup mdc_stat call
The ioctl call inside `mdc_stat` was not checking `errno` and instead
was treating the absolute value of the return code as the error.

Use errno if the return of the ioctl is < 0.

In addition there were some unsafe calls that could error. Make sure
those calls are handled properly and check their `errno`s.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>