The iml-action-dropdown component has a few usecases:
1. A dropdown that takes a single stateful record and produces a list of actions from it.
1. A dropdown that takes a single HSM record and produces a list of actions from it.
1. A dropdown that takes a list of urls, resolves them to stateful records and produces a list of actions.

In addition, this component needs to be used both inside and outside of larger WASM components.

The current coupled nature of all usecases is making it difficult to use this component within a larger WASM component

Instead of having a single component that does everything,
use composition to split out the common parts of the iml-action-dropdown,
and layer the different parts on top.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>