Add initial bootstrap-components module
Fixes #1052.

Add an initial bootstrap-components module that relies solely on
composition of view fns to build up larger fns (as this seems like the
correct model to use with seed). Getting these interfaces right is not
super intuitive, so expect this to evolve over time.

Refactor iml-duration-picker using this module to show viability of this
approach. Also remove iml-popover and replace usage with it's
bootstrap-components equivalent.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>