GH-1056 - Create nginx integration with influx
Fixes #1056.

Description:
Since influxdb provides an http api, the reverse proxy feature of nginx
should be leveraged to allow a query to be run from the application.
This will allow components to be given a stream coming directly from the
database rather than having to go through realtime to get the data. The
end point should be restricted to GET only and should follow
authentication rules. Queries can be made in the form:

https://localhost:8443/influx/<dbname>/<query>

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>