Remove error_for_status call from POST
We are using `error_for_status` to cooerce any unsuccessful status codes
into an `Err`.

However, using this method discards the response body, which may have
useful information in terms of the actual issue with the request /
response.

As a start, this patch will remove `error_for_status` from the top-level
post call to let downstream consumers deal with the error directly.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>