Support both unix socket and TCP communication in the action runner (#1333)
* Support both unix socket and TCP communication in the action runner

The action runner is currently configured to run strictly on a unix
domain socket. This will prevent the docker service from running
correctly. If the proces is being run in systemd then the LISTEN_PID
environment variable will be set and the war server can be spawned with
the unix socket. If the LISTEN_PID environment variable is not defined
it will instead launch the warp server over http (used in docker).

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* Rust formatting

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* - use `get_tcp_or_unix_listener`

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>

* - Run black

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>