Use browser notification API to show alerts
The logic is built around activity health. When the health worsens, a
notification will be shown. When the health gets better a notification will
be updated if shown, otherwise nothing will happen.  In a special case when
no alert remains, a notification will be closed. When an alert that caused
change of the health is known, details of such alert will be shown.

Implementation makes use of a Service Worker, because this is the only way
to get a list of exiting notifications even after the page reloads, closes
and opens.

Update seed because of futures.
Update js_sys because of Array.to_vec().

Signed-off-by: Igor Pashev <pashev.igor@gmail.com>