Enable detail routing
This patch enables us to use a route that contains data (meaning a route
enum variant can now carry some data).

This enables detail pages where we need to include an id in the route.

In addition, this patch renames `Page` enum to `Route`, which seems like
a better name, and extracts it to it's own file.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>