GH-1404 Integrate Influxdb into docker
Fixes #1404.

Description:
Influxdb is a time series database that grafana reads to display metrics using various charts. Influx is currently setup to work with IML using a manager node but is not currently configured to work with a docker setup. This ticket will configure influxdb to work with a docker configuration.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>