Use immutable data structures for warp-drive `Cache`
Fixes #1439.

The [`Cache`](https://github.com/whamcloud/integrated-manager-for-lustre/blob/8352ad767ff7c156168130413b8a91f547650411/iml-wire-types/wire-types.rs#L1982-L1995) is shared between `warp-drive` and `iml-gui` and serves as a way to send a fully populated dataset and then updates afterwards.

It is somewhat difficult to work with within the `iml-gui` seed app, as storing the relevant cache portion to a sub-component's model either means cloning (which is expensive at scale), or passing in a unique id to be stored (`id` in most cases) that implements `Copy`.

Using the later approach is non-trivial to implement, as the sub-component needs to locate the corresponding record by id before using / updating it in the model or view.

Another approach we can take is to use an immutable library such as: https://docs.rs/im/14.1.0.

This will allow for cloning records within sub-components. Since it implements structural sharing, the overhead should be minimal in the read-only case (which is the most typical usecase).

Signed-off-by: Joe Grund <jgrund@whamcloud.io>