Move Vagrantfiles repo into IML repo
Take the repo from https://github.com/whamcloud/Vagrantfiles and plop it
into `integrated-manager-for-lustre/vagrant`.

This will replace https://github.com/whamcloud/Vagrantfiles which will
be marked as deprecated.

Co-locating this repo is the first step of using it to run automated
testing instead of vagrant-libvirt which has been flakey and largely
unmaintained upstream.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>