Use a separate order to remove a record
We currently remove a record inline, but send a message to child
components when one is removed.

We expect the child components to be able to access the record in the
cache before it's deleted, however the child message is async and
happens after the inline delete is processed.

Add a new Msg variant to remove the record. By doing this, we ensure the
remove is scheduled after the chilren can process the delete.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>