Add Series Check to Stragem Library (#1556)
The aggregate_points function assumes that the series will always be in
the influxdb results. I ran into a case in which the series was not in
the results. We should check to make sure that series is not empty
before attempting to get the values and columns from it.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>