Synchronize file processing with incoming streams
If an incoming stream outpaces processing in the mailbox, it will close
before all lines have been flushed to disk.

If this occurs, we may end up trying to process things (ex report /
purge of fids) with an incomplete list.

This patch synchonizes writing all lines to disk such that a connection
will not close until the EOF has been processed and the data is flushed
to disk.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>