Setup a timer to handle stratagem scans at specified intervals (#1445)
* GH-1413 - Setup a timer to handle stratagem scans at specified intervals

Fixes #1413.

Our current solution for running stratagem scans at specied intervals is
to use systemd. Unfortunately, this won't work in the docker
environment. Instead, we should utilize a scheduling container that has
the ability to load a config file with the associated timer data and its
commands. Since the job scheduler will need to generate the timer
config, it will send the config over http to the reverse proxy, where it
will be routed to an http wrapper on the jobber service. Once the
command is received by the http service, it will save the config data
and reload the daemon.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - Update timer endpoints to be in a separate file and only loaded into
the nginx config during the docker nginx setup process

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* fixups

Signed-off-by: Joe Grund <jgrund@whamcloud.io>

* - Fix iml-manager-cli

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - Fix formatting error

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - dedupe the interval for stratagem

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - test to see if removing the rewrite rule works when unconfiguring the
timer

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - Verified that timer configuration and unconfiguration work in single
block

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

Co-authored-by: Joe Grund <grundjoseph@gmail.com>