Use optional environmentfile for config
We have a few config vars that need to be overridden in a docker
environment.

We currently do this by updating the config from print-settings.py
by trying to read an override file

Instead, we can read in an environment file that will be passed into
docker stack and populated at startup.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>