Remove vm cleanup during test setup function
I've run into a case in which the vm cleanup is causing an issue in the
tests. The jenkins cleanup script runs before kicking off the test
suite. We should be able to rely on that cleanup script and not have to
check for running vm's between each test. If we run into issues I will
put this back; therefore I am not going to remove the functions
themselves.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>