Race between agent and manager to resolve action
The `iml-action-runner` service will put a new message onto the queue
before adding it to the internal bookkeeping table of RPCs in flight.

This opens the very small possibility that the RPC can complete before
the action runner can add it to the internal RPC table.

When this occurs, the RPC is never resolved to the caller and ends up
hanging since it looks like like an unknown response to the
action-runner.

Fixes #1870.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>