Use Signed 32 bit integer type for ids
Postgres does not support unsigned numbers (besides the oid type).
As such, all ids in IML are actually signed 32 bit ints.

Update the representation of these types in wire-types and other spots to reflect this.

This will pave the way for using diesel struct types instead of
repeating types in both iml-orm and wire-types.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>