Avoid pg_notify if rows are identical
There are cases (which are becoming more common with upsert + polling)
where rows have not actually changed.

When this occurs we send a pg_notify event, even though the OLD and NEW
rows are identical.

Add a new if branch that will return early if NEW is equal to OLD.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>