Add trait for determining changes between old and new collections
There is a common case where we need to compare a set of old items to a
set of new ones and figure out which ones need to be added / removed or changed.

Instead of writing this logic each time, we can create a trait that will
allow for comparisions and return the changes as `Additions`, `Updates`
and `Deletions`.

This is particularly useful for updating the db with input from daemon
plugins.

General use looks like:

```rust
 let (add, update, remove) = new_items.get_changes(&old_items);
```

Returned items are all borrowed from new and old collections, so no
cloning is needed.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>