Add id column for all SFA db objects
We have a requirement to be able to monitor n SFA appliances at once. In order to do this, we need to have a composite primary key for most SFA objects of the `index` and `storage_system`. Unfortunately, Django does not allow for composite primary or foreign keys even though it's doable in Postgres.

Long term we should move our migrations out of Django and use diesel instead (which allows you to embed migrations into a binary) but for now we'll keep using the Django migrations and add an `unique_together` constraint for `index` and `storage_system` and drop the fk constraints for index on tables.

Additionally, there is a bug where a changed row will appear in both the `Upserts` and `Deletions` of get_changes. We need to have a key fn to determine if something actually belongs in the Deletions.

Add a Identifiable trait that will uniquely id a record so it can be removed from `Deletions` if it's actually a change.

Finally, split out the SFA objects to their own modules within the iml-orm crate for better maintainability.