Update iml-request-retry to support dynamic future creation
There may be a case in which the future factory may need to generate a
different future based on the incremented tries. For example, if there
are two url's that can be fetched for the same information but the first
url fails, we should retry using the second url. If both of the urls are
stored in a vec, then this becomes trivial to fetch using the
incrementing index.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>