Retry FSA Endpoint
There are currently two controllers, each with an endpoint that provide
access to the same api. We try to hit this api using the first
controller, but it's possible that the endpoint will fail. If it does,
it should retry using the second endpoint. If that also fails, then we
can fail overall.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>