Create an iml-config cli tool (#1994)
* Create an iml-config cli tool

Create an iml-config cli tool that will be used to perform config
related tasks. In this patch, a drop-in will be added to nginx that will
load all iml environment variables. These environment variables will
then be used to generate the nginx config file from the chroma template
any time nginx is restarted. This will allow the template variables to
change in the nginx config based on environment variable overloading.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* -cleanups

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - remove nginx setup from service config

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - set_nginx_config is being used by the services test so i'm adding this
back in, but i'm not calling if from the setup function in the service
config.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - Updates

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - CR Changes

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* cleanups

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* Add failing test

Add a failing test case for when there is more than 1 var to be replaced on a line.

Also pass in a HashMap of vars instead of an fn to make testing different cases easier

* - Move nginx config command into existing iml-manager-cli crate and
create a separate binary.
- Fix failing test

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - use iml_tracing

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - Fix clippy errors

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - Tweak the regex a bit

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - Remove set_config_config from service_config.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

* - test1

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

Co-authored-by: Joe Grund <jgrund@whamcloud.io>