Add in_flight handle to active sessions
If a daemon plugin executes for longer than a poll interval it will continue running in the background. This can cause buffering of resources in the case of a misbehaving call.

We've seen this when `lnetctl export` hangs due to large numbers of peers.

We can add a in_flight handle to active sessions.

If we hit the deadline while a plugin is in a poll, we can drop the pending future and poll the plugin again. This will ensure we don't buffer misbehaving resources when a deadline elapses.

Fixes #1998.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>