Do not set environment variables in tests
We have a number of intermittently failing tests that are due to setting environment variables within the tests.

Rust tests run in parallel and the environment is basically a singleton. Therefore, these tests leak state with each other and are non-deterministic.

In addition, we are pulling items between action plugins, where we could factor it out into the env instead.

Fixes #1826.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>