Create an iml-config cli tool (#1988)
Create an iml-config cli tool that will be used to perform config
related tasks. In this patch, a drop-in will be added to nginx that will
load all iml environment variables. These environment variables will
then be used to generate the nginx config file from the chroma template
any time nginx is restarted. This will allow the template variables to
change in the nginx config based on environment variable overloading.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>

Co-authored-by: Joe Grund <jgrund@whamcloud.io>