Create Rust iml-state-machine
- Use `warp-drive` `Cache` as a realtime singleton to get the current
system state. This acts much the same way as the `job_scheduler`
`ObjectCache` does but instead gets realtime updates from the db instead
of needing to be notified of changes by other processes.
- Use petgraph to build a graph consisting of `State` nodes and `Edge`
edges. `Edge` is an enum that is either a `Transition` or a
`Dependency`. Add some methods via an Extenstion trait to find
transitions / shortest transition paths.
- Create a `Job` trait that can either be invoked directly via a
`Command`, or indirectly via a `Transition`.
- Create a `Steps` struct that holds a list of free fns (much like
action plugins). These steps are run serially within a job.
- Refactor service address bindings to not coopt the nginx proxy host.
- Add an input type for `RecordId`s
- Add a http_client to the graphql context
- Add graphql query and mutation for statemachine

Signed-off-by: Joe Grund <jgrund@whamcloud.io>